RentWizAPI/requirements.txt
aniso8601==1.3.0
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
awscli==1.14.24
awsebcli==3.12.1
bcrypt==3.1.3
beautifulsoup4==4.6.0
boto==2.48.0
botocore==1.8.28
cement==2.8.2
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
colorama==0.3.7
config==0.3.9
configparser==3.5.0
constantly==15.1.0
cryptography==2.0
cssselect==1.0.1
decorator==4.1.2
dj-database-url==0.4.2
Django==2.0.1
django-cors-headers==2.1.0
djangorestframework==3.7.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==0.12.2
Flask-And-Redis==0.7
Flask-JSGlue==0.3.1
Flask-Multi-Session==0.1.7
Flask-RESTful==0.3.6
Flask-Session==0.3.1
googlemaps==2.5.1
gunicorn==19.7.1
hyperlink==17.3.0
idna==2.5
incremental==17.5.0
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter-core==4.4.0
lxml==3.8.0
MarkupSafe==1.0
mysqlclient==1.3.10
nbformat==4.4.0
numpy==1.14.0
paramiko==2.2.1
parsel==1.2.0
passlib==1.7.1
pathspec==0.5.0
pickleshare==0.7.4
plotly==2.2.3
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
pyasn1==0.2.3
pyasn1-modules==0.0.11
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyNaCl==1.1.2
pyOpenSSL==17.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
queuelib==1.4.2
redis==2.10.5
requests==2.9.1
rsa==3.4.2
s3transfer==0.1.12
semantic-version==2.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.12
sshtunnel==0.1.2
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
w3lib==1.18.0
wcwidth==0.1.7
websocket-client==0.46.0
Werkzeug==0.12.2
whitenoise==3.3.1
zope.interface==4.4.2
